Add functions for addresses and balance
I went ahead and wrote the functions as a part of
the BitBlockCache class that would allow for
BitBlock to start writing to the address table and
cache address balances.

As of current, it's required that BitBlock cache
all of the block transactions, but it might be
worthwhile to write a function to pull them a la
carte from the RPC.

= CHANGELOG =

- Added fetch_unique_addresses method to
BitBlockCache.

- Added get_address_balance method to BitBlockCache.

- Added get_address_last_tx method to BitBlockCache.