Begin work on BlockReader prototype.
I feel as though going through RPC is much too
slow for the purposes of caching transaction
details. If the user has already downloaded the
entire blockchain, there should be no reason to go
through the RPC. Instead, the user should be able
to access the block files directly. That's what is
hoped for with this prototype.

= CHANGELOG =

- Began work on blockreader.py to implement a
block parser in Python.