Add documentation and refactor a few things
I went through and changed a handful of things, and I commented through the code so that I could understand what I had already done. There's an issue I'm having every now and again where the script_sz is returning an extremely large number when it should be much smaller.